# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FACTA — Independent product intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /collections:
    get:
      operationId: listCollections
      tags: [collections]
      summary: List published curated collections
      parameters:
        - name: goal_slug
          in: query
          schema: { type: ["string", "null"] }
        - name: meal_type
          in: query
          schema: { type: ["string", "null"] }
        - name: retailer_slug
          in: query
          schema: { type: ["string", "null"] }
      responses:
        "200":
          description: Collections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CollectionSummary"
  /collections/{slug}:
    get:
      operationId: getCollection
      tags: [collections]
      summary: Get a collection with its products
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Collection detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Admin — Goals ─────────────────────────────────────────────────────────
...
  /admin/collections:
    post:
      operationId: adminCreateCollection
      tags: [admin]
      summary: Create a curated collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCollectionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionSummary"
  /admin/collections/{id}:
    patch:
      operationId: adminUpdateCollection
      tags: [admin]
      summary: Update a collection
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCollectionInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionSummary"
...
  /admin/pending:
    get:
      operationId: adminListPending
      tags: [admin]
      summary: List pending submissions awaiting review
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Pending submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
  /admin/submissions/{id}/verify:
    post:
      operationId: adminVerifySubmission
      tags: [admin]
      summary: Verify (approve) a pending submission
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewDecision"
      responses:
        "200":
          description: Review recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /admin/submissions/{id}/reject:
    post:
      operationId: adminRejectSubmission
      tags: [admin]
      summary: Reject a pending submission
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewDecision"
      responses:
        "200":
          description: Review recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /admin/products/{id}:
    patch:
      operationId: adminUpdateProduct
      tags: [admin]
      summary: Admin edit product data
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    NutritionFact:
      type: object
      required: [id, productId]
      properties:
        id:               { type: integer }
        productId:        { type: integer }
        servingSize:      { type: ["string", "null"] }
        servingSizeUnit:  { type: ["string", "null"] }
        calories:         { type: ["number", "null"] }
        totalFat:         { type: ["number", "null"] }
        saturatedFat:     { type: ["number", "null"] }
        transFat:         { type: ["number", "null"] }
        sodium:           { type: ["number", "null"] }
        totalCarbs:       { type: ["number", "null"] }
        dietaryFiber:     { type: ["number", "null"] }
        totalSugars:      { type: ["number", "null"] }
        protein:          { type: ["number", "null"] }
    IngredientItem:
      type: object
      required: [id, name]
      properties:
        id:             { type: integer }
        name:           { type: string }
        nameZh:         { type: ["string", "null"] }
        riskLevel:      { type: ["string", "null"] }
        riskReason:     { type: ["string", "null"] }
        evidenceStrength: { type: ["string", "null"] }
        isAdditive:     { type: boolean }
    AllergenAlert:
      type: object
      required: [name, severity]
      properties:
        name:     { type: string }
        nameZh:   { type: ["string", "null"] }
        severity: { type: string }
        source:   { type: ["string", "null"] }
    ProductSummary:
      type: object
      required: [id, name, verificationStatus]
      properties:
        id:                 { type: integer }
        name:               { type: string }
        nameZh:             { type: ["string", "null"] }
        brandName:          { type: ["string", "null"] }
        imageUrl:           { type: ["string", "null"] }
        categorySlug:       { type: ["string", "null"] }
        categoryName:       { type: ["string", "null"] }
        verificationStatus: { type: string }
        overallScore:       { type: ["integer", "null"] }
        scoreGrade:         { type: ["string", "null"] }
        barcode:            { type: ["string", "null"] }
        retailerName:       { type: ["string", "null"] }
        priceNtd:           { type: ["number", "null"] }
    Product:
      type: object
      required: [id, name, verificationStatus]
      properties:
        id:                 { type: integer }
        name:               { type: string }
        nameZh:             { type: ["string", "null"] }
        brandName:          { type: ["string", "null"] }
        imageUrl:           { type: ["string", "null"] }
        barcode:            { type: ["string", "null"] }
        categorySlug:       { type: ["string", "null"] }
        categoryName:       { type: ["string", "null"] }
        categoryNameZh:     { type: ["string", "null"] }
        verificationStatus: { type: string }
        dataCompleteness:   { type: ["number", "null"] }
        retailerName:       { type: ["string", "null"] }
        retailerId:         { type: ["integer", "null"] }
        priceNtd:           { type: ["number", "null"] }
        netWeight:          { type: ["string", "null"] }
        ingredientsList:    { type: ["string", "null"] }
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/IngredientItem"
        nutritionFacts:
          oneOf:
            - $ref: "#/components/schemas/NutritionFact"
            - type: "null"
        allergens:
          type: array
          items:
            $ref: "#/components/schemas/AllergenAlert"
        createdAt:  { type: string }
        updatedAt:  { type: string }

    # ── Evaluation / FACTA Report ─────────────────────────────────────────
...
    Submission:
      type: object
      required: [id, productName, status, createdAt]
      properties:
        id:              { type: integer }
        productName:     { type: string }
        brandName:       { type: ["string", "null"] }
        barcode:         { type: ["string", "null"] }
        status:          { type: string }
        ocrStatus:       { type: ["string", "null"] }
        extractedIngredients: { type: ["string", "null"] }
        provisionalScore: { type: ["integer", "null"] }
        provisionalGrade: { type: ["string", "null"] }
        dataCompleteness: { type: ["number", "null"] }
        userSession:     { type: ["string", "null"] }
        reviewNote:      { type: ["string", "null"] }
        createdAt:       { type: string }
        updatedAt:       { type: string }

    # ── User Preferences ──────────────────────────────────────────────────
...
    CollectionSummary:
      type: object
      required: [id, slug, nameZh, status]
      properties:
        id:            { type: integer }
        slug:          { type: string }
        name:          { type: string }
        nameZh:        { type: string }
        descriptionZh: { type: ["string", "null"] }
        goalId:        { type: ["integer", "null"] }
        retailerId:    { type: ["integer", "null"] }
        mealType:      { type: ["string", "null"] }
        status:        { type: string }
        productCount:  { type: integer }
    CollectionDetail:
      allOf:
        - $ref: "#/components/schemas/CollectionSummary"
        - type: object
          properties:
            products:
              type: array
              items:
                type: object
                properties:
                  product:
                    $ref: "#/components/schemas/ProductSummary"
                  reason:    { type: ["string", "null"] }
                  reasonZh:  { type: ["string", "null"] }

    # ── Admin input schemas ────────────────────────────────────────────────
...
    AdminCollectionInput:
      type: object
      properties:
        slug:          { type: string }
        name:          { type: string }
        nameZh:        { type: string }
        descriptionZh: { type: ["string", "null"] }
        goalId:        { type: ["integer", "null"] }
        retailerId:    { type: ["integer", "null"] }
        mealType:      { type: string }
        status:        { type: string }
        productIds:    { type: array, items: { type: integer } }

    # ── Admin ─────────────────────────────────────────────────────────────
    ReviewDecision:
      type: object
      required: [reviewedBy]
      properties:
        reviewedBy: { type: string }
        note:       { type: ["string", "null"] }
    ProductUpdate:
      type: object
      properties:
        name:               { type: string }
        nameZh:             { type: ["string", "null"] }
        ingredientsList:    { type: ["string", "null"] }
        verificationStatus: { type: ["string", "null"] }
        imageUrl:           { type: ["string", "null"] }
        priceNtd:           { type: ["number", "null"] }
        netWeight:          { type: ["string", "null"] }